requirements.txt
django<1.12
django-phonenumber-field
django-allauth
django-summernote
Pillow
raven
boto3
django-storages
djangorestframework
django-filter
psycopg2

